sliding-window-logs/requirements.txt
async-timeout==4.0.3
blinker==1.7.0
click==8.1.7
Flask==3.0.2
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
redis==5.0.2
Werkzeug==3.0.1
zipp==3.17.0
